wip: break out filter components
