+++ package.json
@@ -32,6 +32,7 @@
 		"react-window": "^2.2.7",
 		"sanitize-html": "^2.17.4",
 		"stemmer": "^2.0.1",
+		"virtua": "^0.49.1",
 		"wouter": "^3.9.0",
 		"zustand": "^5.0.13"
 	},
