+++ src/pages/SearchPage.tsx
@@ -16,7 +16,7 @@ export default function SearchPage() {
 		<Root>
 			<div className="flex flex-col gap-3 p-4" style={{ height: "calc(100svh - 60px)" }}>
 				<div className="flex flex-col gap-3">
-					<QuestionListHeader results={questions.length} />
+					{/* <QuestionListHeader results={questions.length} /> */}
 					<SearchInput />
 					<SearchOptions />
 				</div>
